refactor: util.js에서 dom관련 dom.js로 분리
